core/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
Pillow==10.1.0
psycopg2-binary==2.9.9
Pygments==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytils==0.4.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
